pyproject.toml
[project]
name = "vtii"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.78.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "ruff>=0.11.9",
    "scipy>=1.13.1",
    "sounddevice>=0.5.1",
    "tqdm>=4.67.1",
    "websockets>=15.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
